+++ src/init.c
@@ -4,7 +4,7 @@
 #include <Rinternals.h>
 
 
-extern SEXP get_all_body_info_();
+extern SEXP get_body_state_();
 
 extern SEXP cpVect_();
 extern SEXP cpVect_to_list_();
@@ -101,7 +101,7 @@ extern SEXP cpTransformRigid_();
 
 static const R_CallMethodDef CEntries[] = {
 
-  {"get_all_body_info_", (DL_FUNC) &get_all_body_info_, 1},
+  {"get_body_state_", (DL_FUNC) &get_body_state_, 1},
 
   //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   // Structs