+++ NEWS.md
@@ -1,5 +1,11 @@
 
 
+# chipmunkcore 0.1.4
+
+* Add `get_body_state(bodies)` which will create a data.frame with the current
+state of each body in the given list of `bodies`. i.e. position, velocity, 
+force, angle, angular velocity and torque.
+
 # chipmunkcore 0.1.3
 
 * API: move R6 higher-level wrapper to its own package [chipmunkbasic](https://github.com/coolbutuseless/chipmunkbasic)